feat: fix GCal events padding on mobile
